house_of_minds/requirements_dolphin.txt
# Dolphin Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiohttp==3.9.1
pydantic==2.5.0
python-dotenv==1.0.0

# House of Minds components (from existing requirements)
textblob==0.17.1
scikit-learn==1.3.2
tinydb==4.8.0
python-multipart==0.0.6

# Additional utilities
httpx==0.25.2
asyncio-compat==0.1.2
